java file server with white/black list
Please note that
If you want to use the white list feature, you need to create a file called "whiteList.wl" and place it in the directory as the start-up jar package file path. In the "whiteList.wl" file, fill in an accessible IP address per line.
The blacklist is similar. the different is blacklist file called "blackList.wl"
But blacklists have a higher priority than white lists, which means that even if the IP contained in the blacklist exists in the whitelist, the blacklist policy will be used to deny current IP access.
The file server supports three startup parameters, the first is the service port, the second is the file root directory displayed outside, and the third is the white/black list directory.